feat(dictionary-category): 实现词典与分类的关联功能
- 新增 CategoryRelativeRequest 类用于处理分类关联请求
- 修改 DictionaryCategory 实体类，将 dictionary_id 类型从 Integer 改为 Long
- 更新数据库表结构，将 dictionary_category 表的 dictionary_id 类型改为 bigint
- 在 DictionaryCategoryController 中添加分类关联相关接口
- 在 DictionaryCategoryService 接口中定义新的关联分类方法
- 实现 DictionaryCategoryImpl 中的分类关联逻辑，包括批量添加和查询功能

+++ sql/create_table.sql
@@ -172,7 +172,7 @@ CREATE TABLE category (
 
 CREATE TABLE dictionary_category (
     id INT UNSIGNED AUTO_INCREMENT NOT NULL, -- 新增自增主键列
-    dictionary_id INT UNSIGNED NOT NULL,
+    dictionary_id bigint UNSIGNED NOT NULL,
     category_id INT UNSIGNED NOT NULL,
     sort_order INT DEFAULT 0 COMMENT '分类内排序',
     created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
